refactor: extract PartClassificationState from JsonXModelLoader
